requirements.txt
amqp==2.4.2
billiard==3.5.0.5
celery==4.2.2
Django==2.1.7
entrypoints==0.3
flake8==3.7.7
kombu==4.3.0
marshmallow==3.2.1
mccabe==0.6.1
pika==1.0.1
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pyflakes==2.1.1
python-dotenv==0.10.3
pytz==2018.9
six==1.12.0
vine==1.3.0
youtube-dl==2019.9.28
